requirements.txt
pymongo
mysql-connector-python
tqdm